클라이언트에 보낸 카운트 로그 찍기

+++ ServerCore/PacketBuffer.cpp
@@ -9,7 +9,7 @@ PacketBuffer::PacketBuffer(int32 bufferSize)
     m_Buffer.resize(m_Capacity);
 }
 
-bool PacketBuffer::OnWrite(int32 size)
+bool PacketBuffer::OnWrite(size_t size)
 {
     if (size > FreeSize())
         return false;
